+++ src/main/java/Duke/Deadline.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
