+++ src/main/java/Duke/Event.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class Event extends Task {
 
     protected String from;
