+++ src/main/java/Duke/Task.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class Task {
     protected String description;
     protected boolean isDone;
